requirements_temp.txt
transformers==4.50.0
deepspeed==0.16.4
triton==3.0.0
accelerate==1.4.0
torchcodec==0.2
